remove all VS related crap | reorganize files
