--- Allegro 2048.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30517.126
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Allegro 2048", "Allegro 2048.vcxproj", "{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Debug|x64.ActiveCfg = Debug|x64
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Debug|x64.Build.0 = Debug|x64
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Debug|x86.ActiveCfg = Debug|Win32
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Debug|x86.Build.0 = Debug|Win32
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Release|x64.ActiveCfg = Release|x64
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Release|x64.Build.0 = Release|x64
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Release|x86.ActiveCfg = Release|Win32
-		{E4FE8BA3-FF60-478F-9E11-7E0F49F62BC0}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5ECA9CF0-1A87-4272-8566-020446BB4AF2}
-	EndGlobalSection
-EndGlobal
