--- packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages xmlns="urn:packages">
-  <package id="Allegro" version="5.2.6" targetFramework="native" />
-  <package id="AllegroDeps" version="1.11.0" targetFramework="native" />
-</packages>
